extract shared functions to module
