+++ app/src/lib/registry.ts
@@ -1,55 +1,7 @@
-import * as Path from 'path'
 import * as ChildProcess from 'child_process'
-
-// TODO: extract this so it's shared with squirrel-updater
-function getPowerShellPath(): string {
-  const systemRoot = process.env['SystemRoot']
-  if (systemRoot) {
-    const system32Path = Path.join(process.env.SystemRoot, 'System32')
-    return Path.join(
-      system32Path,
-      'WindowsPowerShell',
-      'v1.0',
-      'powershell.exe'
-    )
-  } else {
-    return 'powershell.exe'
-  }
-}
+import { spawn as spawnPowerShell } from './powershell'
 
 // TODO: this is also shared with squirrel-updater - can we unify this?
-
-/** Spawn a command with arguments and capture its output. */
-function spawn(command: string, args: ReadonlyArray<string>): Promise<string> {
-  try {
-    const child = ChildProcess.spawn(command, args as string[])
-    return new Promise<string>((resolve, reject) => {
-      let stdout = ''
-      child.stdout.on('data', data => {
-        stdout += data
-      })
-
-      child.on('close', code => {
-        if (code === 0) {
-          resolve(stdout)
-        } else {
-          reject(new Error(`Command "${command} ${args}" failed: "${stdout}"`))
-        }
-      })
-
-      child.on('error', (err: Error) => {
-        reject(err)
-      })
-
-      // This is necessary if using Powershell 2 on Windows 7 to get the events
-      // to raise.
-      // See http://stackoverflow.com/questions/9155289/calling-powershell-from-nodejs
-      child.stdin.end()
-    })
-  } catch (error) {
-    return Promise.reject(error)
-  }
-}
 
 export interface IRegistryEntry {
   readonly name: string
@@ -101,7 +53,7 @@ export async function readRegistryKeySafe(
 
   const results = new Array<IRegistryEntry>()
 
-  const stdout = await spawn(getPowerShellPath(), args)
+  const stdout = await spawnPowerShell(args)
   const jsonText = stdout.trim()
 
   if (jsonText.length) {
